[geoclue-provider-mlsdb] Allow country-specific datafiles. Contributes to JB#34561
This commit allows country or region specific data files to be
installed to separate data directories.

At the moment, the plugin still only loads a single data file into
memory, however in the future it should be improved to load further
data files as required until it finds the cell id.

Contributes to JB#34561