[geoclue-provider-mlsdb] Enable by default if positioning is enabled. Contributes to JB#34561
This commit makes the MLSDB positioning plugin assume that it is
enabled by default if the offline cell positioning setting has not
yet been set to either true or false, if positioning is enabled.

If the user has explicitly disabled offline cell positioning, or
positioning in general, then the MLSDB positioning plugin will still
be disabled.

Contributes to JB#34561