[geoclue-provider-mlsdb] Detect changes to location.conf. Contributes to JB#34743
This commit adds a QFileSystemWatcher to watch the location.conf file
for changes, so that the behaviour of the plugin dynamically reacts
to the user changing their location settings.

Upon being disabled while running, its positioning updates will become
disabled, and if not re-enabled within the QuitIdleTime (30s), the
plugin will exit.

Contributes to JB#34743