[geoclue-provider-mlsdb] Use location code in lookups. Contributes to JB#34744
Previously, the plugin assumed that the cellId was globally unique.
This is not true, instead the cellId is only unique within a given
location code (LAC on GSM or TAC on LTE networks).

This commit ensures that the key for location lookups is composed
of the location area code and the cell id.

It also regenerates the databases so that they contain the appropriate
data.

Contributes to JB#34744