[geoclue-provider-mlsdb] Listen to config directory changes. Contributes to JB#36857
When QSettings writes values to the settings file, it can perform
an atomic replacement via delete+rename.  This can result in missed
change signals if the QFileSystemWatcher is not watching the directory
but just the file.

This commit ensures that we listen to config directory changes as
well as config file changes.

Contributes to JB#36857