Changed the algorithm and query parameters in the Mozilla service
If there are no cellular towers and WiFi modems, then the request to Mozilla service does not send. The service must send complete information about each cell tower. If there is no modem wifi signal, do not send this modem to the service. In fallbacks, lacf and ipf must be set to false, otherwise the service sends often incorrect coordinates.