[geoclue-provider-mlsdb] Reduce position calculation frequency. Contributes to JB#41844
In some phone models, the cellChanged signal is frequently triggered.
Accordingly, this leads to a large number of requests and increased
energy consumption.

By default, the timer is triggered once every 10 seconds, which is
enough for inaccurate positioning.

In this commit, we are reducing the number of unnecessary conversions
and calculations.