[geoclue-provider-mlsdb] Change location algorithm. Contributes to JB#41844
In some phone models, the cellChanged signal is frequently triggered. Accordingly, this leads to a large number of requests and increased energy consumption.
By default, the timer is triggered once every 10 seconds, which is enough to inaccurate positioning

In this case, we are reducing the number of unnecessary conversions and calculations