[geoclue-provider-mlsdb] Cooperatively observe allowed data sources. Contributes to JB#33753
The MDM API now supports explicitly enabling or disabling various
data sources which may be used by location plugins when performing
position determination.

This commit ensures that the MLSDB plugin cooperates with those
settings.

In the future, fine-grained access control and proper daemonised
API services will obviate the need for this PR to a large extent.